+++ scripts/macros/spells/findFamiliar.js
@@ -47,7 +47,10 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
             'chris-premades': {
                 'spell': {
                     'findFamiliar': true
-                }
+                },
+                'vae': {
+                    'button': attackData.name
+                },
             }
         }
     }